+++ examples/input.rs
@@ -6,12 +6,10 @@ struct InputApp;
 
 impl TextualApp for InputApp {
     fn compose(&mut self) -> AppRoot {
-        AppRoot::new().with_child(
-            Container::new().with_compose(compose![
-                Input::new().with_placeholder("First Name"),
-                Input::new().with_placeholder("Last Name"),
-            ]),
-        )
+        AppRoot::new().with_child(Container::new().with_compose(compose![
+            Input::new().with_placeholder("First Name"),
+            Input::new().with_placeholder("Last Name"),
+        ]))
     }
 }