+++ examples/input_types.rs
@@ -6,16 +6,14 @@ struct InputTypesApp;
 
 impl TextualApp for InputTypesApp {
     fn compose(&mut self) -> AppRoot {
-        AppRoot::new().with_child(
-            Container::new().with_compose(compose![
+        AppRoot::new().with_child(Container::new().with_compose(compose![
                 Input::new()
                     .with_placeholder("An integer")
                     .with_type(InputType::Integer),
                 Input::new()
                     .with_placeholder("A number")
                     .with_type(InputType::Number),
-            ]),
-        )
+            ]))
     }
 }